+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes relevant to the users of python-tcod are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/) since version `2.0.0`.
 
 ## [Unreleased]
+
+## [13.6.0] - 2022-02-19
 ### Added
 - `BasicMixer` and `Channel` classes added to `tcod.sdl.audio`.  These handle simple audio mixing.
 - `AudioDevice.convert` added to handle simple conversions to the active devices format.